requirements.txt
# 🔧 Core Development Tools
pip==24.0
setuptools==65.0.0
wheel==0.40.0

# 🎨 CLI & Interface
rich==13.7.0
typer==0.9.0
prompt_toolkit==3.0.36  # Adjusted for compatibility
colorama==0.4.6

# 📊 Data Processing & Analysis
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.8.2

# 🔍 System Information
psutil==5.9.8
distro==1.9.0

# 🔒 Security & Environment
python-dotenv==1.0.1
cryptography==42.0.5
keyring==24.3.1

# 🌐 Google API & Cloud
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
gspread==6.0.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0

# 🧪 Testing & Quality Assurance
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.2.0
flake8==7.0.0
mypy==1.9.0
isort==5.13.2
bandit==1.7.7

# 📝 Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3
mkdocs-material==9.5.11

# 🔄 Version Control & Development
pre-commit==3.6.2

# 📦 Build & Distribution
build==1.0.3
twine==5.0.0

# 🔍 Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.5